feat: refactor file structure and add kyc + token-price fetching examples
