+++ hybrid-compute/offchain/offchain.py
@@ -1,0 +1,40 @@
+import os
+from web3 import Web3
+from jsonrpclib.SimpleJSONRPCServer import SimpleJSONRPCServer, SimpleJSONRPCRequestHandler
+
+from add_sub_2.add_sub_2_offchain import offchain_addsub2
+from ramble.ramble_offchain import offchain_ramble
+from check_kyc.check_kyc_offchain import offchain_checkkyc
+from get_token_price.get_token_price_offchain import offchain_getprice
+
+PORT = int(os.environ['OC_LISTEN_PORT'])
+assert (PORT != 0)
+
+
+
+
+def selector(name):
+    nameHash = Web3.to_hex(Web3.keccak(text=name))
+    return nameHash[2:10]
+
+
+class RequestHandler(SimpleJSONRPCRequestHandler):
+    rpc_paths = ('/', '/hc')
+
+
+def server_loop():
+    server = SimpleJSONRPCServer(
+        ('0.0.0.0', PORT), requestHandler=RequestHandler)
+    server.register_function(offchain_addsub2, selector(
+        "addsub2(uint32,uint32)"))  # 97e0d7ba
+    server.register_function(
+        offchain_ramble,  selector("ramble(uint256,bool)"))
+    server.register_function(
+        offchain_checkkyc, selector("checkkyc(string)"))
+    server.register_function(
+        offchain_getprice, selector("getprice(string)"))
+
+    server.serve_forever()
+
+
+server_loop()  # Run until killed
